[HBASE-10699] Reduce unnecessary object allocation related to ArrayList [HBASE-10699] Fix the building problem. [HBASE-10698] Updating the HBase.thrift file to work with Thrift2 C++ client [HBASE-10709] ClientScanner with prefetching [HBASE-9930] Fix TestPerRequestProfiling [HBASE-8500] Provide a drain only option to rolling_restart [master] Add a param for hbck to decide whether to clear meta entry for unassigned region
Summary: Multiple Commits

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1576909 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/regionserver/RegionOverloadedException.java
@@ -29,7 +29,7 @@ public class RegionOverloadedException extends RegionException {
    * @return the new exception with complete information
    */
   public static RegionOverloadedException create(RegionOverloadedException roe,
-      List<Throwable> exceptions, int waitMillis) {
+      List<Throwable> exceptions, long waitMillis) {
     StringBuilder sb = new StringBuilder(roe.getMessage());
     for (Throwable t : exceptions) {
       if (t != roe) {
